+++ pelix/remote/edef_io.py
@@ -9,7 +9,7 @@ specifications, section 122.8.
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.1
+:version: 3.2.2
 
 ..
 
@@ -35,11 +35,12 @@ from typing import Any, Dict, Iterable, List, Optional, Tuple, Union
 import pelix.constants
 import pelix.remote
 from pelix.remote.beans import EndpointDescription
+from pelix.utilities import check_xml_no_doctype
 
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (3, 2, 1)
+__version_info__ = (3, 2, 2)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format
@@ -212,7 +213,11 @@ class EDEFReader:
 
         :param xml_str: An XML string
         :return: The list of parsed EndpointDescription
+        :raise ValueError: Invalid XML document
         """
+        # Refuse documents declaring entities before giving them to the parser
+        check_xml_no_doctype(xml_str)
+
         # Parse the document
         root: ElementTree.Element = ElementTree.fromstring(xml_str)
         if root.tag != TAG_ENDPOINT_DESCRIPTIONS: